Changes
1. Made a "BodySourceManager" prefab, a "PlayerBody" prefab to enable
Kinect interaction and added in the "DetectJoints.cs" script 2. Edited
the "TouchToSwitchScene.cs" and made a "Quit Game Manager" prefab to
quit the game via keyboard and touching the quitting ball; 3. Edited the
"BallData.cs" to make the answer balls recognize the "PlayerBody"; 4.
Added in new sprites; 5. Edited the scenes and adjusted the camera in
each scene according to the Kinect Scene.